omr_app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Image processing
opencv-python-headless==4.10.0.84
numpy==2.2.1
Pillow==11.1.0

# HTTP
requests==2.32.3
httpx==0.28.1

# Firebase
firebase-admin==6.7.0

# Excel export
openpyxl==3.1.5

# PDF template generation
fpdf2==2.8.2

# Reporting / statistics
scipy==1.15.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Rate limiting
slowapi==0.1.9

# Monitoring
sentry-sdk[fastapi]==2.19.2
